+++ CHANGES.md
@@ -1,5 +1,14 @@
 # Changes
 
+## cordova-sqlite-storage 2.1.3
+
+##### cordova-sqlite-legacy-core 1.0.5
+
+###### cordova-sqlite-legacy-express-core 1.0.3
+
+- Resolve Java 6/7/8 concurrent map compatibility issue reported in litehelpers/Cordova-sqlite-storage#726, THANKS to pointer by @NeoLSN (Jason Yang/楊朝傑) in litehelpers/Cordova-sqlite-storage#727.
+- selfTest database cleanup do not ignore close or delete error on any platforms
+
 ## cordova-sqlite-storage 2.1.2
 
 ##### cordova-sqlite-legacy-core 1.0.4